# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hostack API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/build-rules:
    get:
      operationId: listBuildRules
      tags: [build-rules]
      summary: List build rules for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of build rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildRuleListResponse"
        "401":
          description: Unauthorized
        "404":
          description: Project not found
    post:
      operationId: createBuildRule
      tags: [build-rules]
      summary: Create a build rule
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBuildRuleBody"
      responses:
        "201":
          description: Build rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildRule"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Project not found
  /projects/{projectId}/build-rules/{ruleId}:
    put:
      operationId: updateBuildRule
      tags: [build-rules]
      summary: Update a build rule
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBuildRuleBody"
      responses:
        "200":
          description: Updated build rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildRule"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteBuildRule
      tags: [build-rules]
      summary: Delete a build rule
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /projects/{projectId}/deploy-webhooks:
    get:
      operationId: listDeployWebhooks
      tags: [deploy-webhooks]
      summary: List deploy webhooks for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of deploy webhooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployWebhookListResponse"
        "401":
          description: Unauthorized
        "404":
          description: Project not found
    post:
      operationId: createDeployWebhook
      tags: [deploy-webhooks]
      summary: Generate a new deploy webhook
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeployWebhookBody"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployWebhook"
        "401":
          description: Unauthorized
        "404":
          description: Project not found
  /projects/{projectId}/deploy-webhooks/{webhookId}:
    delete:
      operationId: deleteDeployWebhook
      tags: [deploy-webhooks]
      summary: Delete a deploy webhook
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /webhooks/{projectId}/{secret}:
    post:
      operationId: triggerDeployWebhook
      tags: [deploy-webhooks]
      summary: Trigger a deployment via webhook (public, no auth required)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: secret
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                branch:
                  type: string
                  description: Branch name to deploy. Defaults to "main" if not provided.
      responses:
        "200":
          description: Webhook received but deployment was skipped (e.g. auto-deploy disabled)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
        "201":
          description: Deployment triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment:
                    $ref: "#/components/schemas/Deployment"
                required:
                  - deployment
        "404":
          description: Invalid webhook
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Deployment:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: string
          nullable: true
        status:
          type: string
          enum: [queued, preparing, cloning, detecting, installing, building, uploading, ready, deployed, failed]
        environment:
          type: string
          enum: [production, preview]
        triggerType:
          type: string
          enum: [manual, webhook, api, rollback]
          nullable: true
        branch:
          type: string
          nullable: true
        commitHash:
          type: string
          nullable: true
        commitMessage:
          type: string
          nullable: true
        deploymentUrl:
          type: string
          nullable: true
        outputDirectory:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        durationSeconds:
          type: number
          nullable: true
        prNumber:
          type: number
          nullable: true
      required:
        - id
        - projectId
        - status
        - environment
        - createdAt
...
    BuildRule:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        branchPattern:
          type: string
        environment:
          type: string
          enum: [production, preview]
        autoDeploy:
          type: boolean
        buildCommandOverride:
          type: string
          nullable: true
        installCommandOverride:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - branchPattern
        - environment
        - autoDeploy
        - createdAt
        - updatedAt
    BuildRuleListResponse:
      type: object
      properties:
        buildRules:
          type: array
          items:
            $ref: "#/components/schemas/BuildRule"
      required:
        - buildRules
    CreateBuildRuleBody:
      type: object
      properties:
        branchPattern:
          type: string
        environment:
          type: string
          enum: [production, preview]
        autoDeploy:
          type: boolean
        buildCommandOverride:
          type: string
        installCommandOverride:
          type: string
      required:
        - branchPattern
    UpdateBuildRuleBody:
      type: object
      properties:
        branchPattern:
          type: string
        environment:
          type: string
          enum: [production, preview]
        autoDeploy:
          type: boolean
        buildCommandOverride:
          type: string
        installCommandOverride:
          type: string
    DeployWebhook:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        secret:
          type: string
        label:
          type: string
        lastTriggeredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - secret
        - label
        - createdAt
    DeployWebhookListResponse:
      type: object
      properties:
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/DeployWebhook"
      required:
        - webhooks
    CreateDeployWebhookBody:
      type: object
      properties:
        label:
          type: string
...